+++ services/resourcemanager/CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.7.8 (2024-04-09)
+
+- Set config.ContextHTTPRequest in Execute method
+
 ## v0.7.7 (2024-02-28)
 
 - Update `core` to [`v0.10.0`](../../core/CHANGELOG.md#v0100-2024-02-27)